+++ README.md
@@ -1,1 +1,4 @@
 # boilerplate-es7-typescript
+
+---root---src--|
+               |
