--- jestSetup.ts
@@ -1,1 +1,0 @@
-global.rootDir = process.cwd()
