+++ packages/server/.eslintrc.js
