+++ packages/server/README.md
@@ -1,0 +1,1 @@
+# boilerplate-es7-typescript
