+++ packages/server/src/app/helpers/CustomError.ts
