+++ packages/server/src/configs/index.ts
