+++ packages/server/src/test/models/Home.spec.ts
@@ -1,15 +1,15 @@
 'use strict'
 
-import { HealthModel } from '../../Domain: Home/models/'
+import { VersionModel } from '../../app/models/'
 
 describe('Home Module Testing', () => {
   test('Get Server Version', async() => {
     const expectedResult = {
       name: 'nodejs-es7-boilerplate',
       version: '1.0.0',
-      description: 'ES7 boilerplate',
+      description: 'Typescript boilerplate ES7',
     }
-    let result = await HealthModel.get();
+    let result = await VersionModel.get();
     expect(result).toMatchObject(expectedResult);
   })
 })
